mongo: fix MongoContainers for latest version of Docker

+++ grails-data-mongodb/core/src/test/groovy/org/apache/grails/data/mongo/core/GrailsDataMongoTckManager.groovy
@@ -66,28 +66,11 @@ class GrailsDataMongoTckManager extends GrailsDataTckManager {
     MongoDatastore multiTenantMultiDataSourceDatastore
 
     @Override
-    void setupSpec() {
-        super.setupSpec()
-        mongoDBContainer = new MongoDBContainer(AbstractMongoGrailsExtension.desiredMongoDockerName)
-        mongoDBContainer.start()
-        mongoDBContainer.followOutput(new Slf4jLogConsumer(LoggerFactory.getLogger("testcontainers")))
-
-        configuration = [
-                (MongoSettings.SETTING_DATABASE_NAME): 'test',
-                (MongoSettings.SETTING_HOST)         : mongoDBContainer.host,
-                (MongoSettings.SETTING_PORT)         : mongoDBContainer.getMappedPort(AbstractMongoGrailsExtension.DEFAULT_MONGO_PORT) as String,
-                //TODO: 'grails.mongodb.url': "mongodb://${host}:${port as String}/myDb" as String
-        ]
-    }
-
-    @Override
-    void cleanupSpec() {
-        super.cleanupSpec()
-        mongoDBContainer.stop()
-    }
-
-    @Override
     Session createSession() {
+        if (mongoDBContainer == null || !mongoDBContainer.isRunning()) {
+            return null
+        }
+        System.setProperty('mongodb.gorm.suite', 'true')
         def allClasses = getDomainClasses() as Class[]
         def ctx = new GenericApplicationContext()
         ctx.refresh()
@@ -136,17 +119,19 @@ class GrailsDataMongoTckManager extends GrailsDataTckManager {
 
     @Override
     void destroy() {
-        mongoDatastore.getMongoClient().listDatabaseNames().findAll {!(it in ['admin', 'config', 'local']) }.each {
-            try {
-                mongoDatastore.getMongoClient().getDatabase(it).drop()
+        if (mongoDatastore != null) {
+            mongoDatastore.getMongoClient().listDatabaseNames().findAll {!(it in ['admin', 'config', 'local']) }.each {
+                try {
+                    mongoDatastore.getMongoClient().getDatabase(it).drop()
+                }
+                catch(e) {
+                    log.warn("Could not drop ${it}")
+                }
             }
-            catch(e) {
-                log.warn("Could not drop ${it}")
+            mongoDatastore.buildIndex()
+            for (cls in getDomainClasses()) {
+                GormEnhancer.findValidationApi(cls).setValidator(null)
             }
-        }
-        mongoDatastore.buildIndex()
-        for (cls in getDomainClasses()) {
-            GormEnhancer.findValidationApi(cls).setValidator(null)
         }
 
         super.destroy()
@@ -154,20 +139,22 @@ class GrailsDataMongoTckManager extends GrailsDataTckManager {
 
     @Override
     boolean supportsMultipleDataSources() {
-        true
+        mongoDBContainer != null && mongoDBContainer.isRunning()
     }
 
     @Override
     void setupMultiDataSource(Class... domainClasses) {
-        String host = mongoDBContainer.host
-        int port = mongoDBContainer.getMappedPort(AbstractMongoGrailsExtension.DEFAULT_MONGO_PORT)
-        Map config = [
-                'grails.mongodb.url'       : "mongodb://${host}:${port}/tckDefaultDB" as String,
-                'grails.mongodb.connections': [
-                        'secondary': ['url': "mongodb://${host}:${port}/tckSecondaryDB" as String],
-                ],
-        ]
-        multiDataSourceDatastore = new MongoDatastore(DatastoreUtils.createPropertyResolver(config), domainClasses)
+        if (mongoDBContainer != null && mongoDBContainer.isRunning()) {
+            String host = mongoDBContainer.host
+            int port = mongoDBContainer.getMappedPort(AbstractMongoGrailsExtension.DEFAULT_MONGO_PORT)
+            Map config = [
+                    'grails.mongodb.url'       : "mongodb://${host}:${port}/tckDefaultDB" as String,
+                    'grails.mongodb.connections': [
+                            'secondary': ['url': "mongodb://${host}:${port}/tckSecondaryDB" as String],
+                    ],
+            ]
+            multiDataSourceDatastore = new MongoDatastore(DatastoreUtils.createPropertyResolver(config), domainClasses)
+        }
     }
 
     @Override
@@ -184,30 +171,32 @@ class GrailsDataMongoTckManager extends GrailsDataTckManager {
     @Override
     def getServiceForConnection(Class serviceType, String connectionName) {
         multiDataSourceDatastore
-                .getDatastoreForConnection(connectionName)
-                .getService(serviceType)
+                ?.getDatastoreForConnection(connectionName)
+                ?.getService(serviceType)
     }
 
     @Override
     boolean supportsMultiTenantMultiDataSource() {
-        true
+        mongoDBContainer != null && mongoDBContainer.isRunning()
     }
 
     @Override
     void setupMultiTenantMultiDataSource(Class... domainClasses) {
-        String host = mongoDBContainer.host
-        int port = mongoDBContainer.getMappedPort(AbstractMongoGrailsExtension.DEFAULT_MONGO_PORT)
-        Map config = [
-                'grails.gorm.multiTenancy.mode'               : MultiTenancySettings.MultiTenancyMode.DISCRIMINATOR,
-                'grails.gorm.multiTenancy.tenantResolverClass' : SystemPropertyTenantResolver,
-                'grails.mongodb.url'                           : "mongodb://${host}:${port}/tckMtDefaultDB" as String,
-                'grails.mongodb.connections'                   : [
-                        'secondary': ['url': "mongodb://${host}:${port}/tckMtSecondaryDB" as String],
-                ],
-        ]
-        multiTenantMultiDataSourceDatastore = new MongoDatastore(
-                DatastoreUtils.createPropertyResolver(config), domainClasses
-        )
+        if (mongoDBContainer != null && mongoDBContainer.isRunning()) {
+            String host = mongoDBContainer.host
+            int port = mongoDBContainer.getMappedPort(AbstractMongoGrailsExtension.DEFAULT_MONGO_PORT)
+            Map config = [
+                    'grails.gorm.multiTenancy.mode'               : MultiTenancySettings.MultiTenancyMode.DISCRIMINATOR,
+                    'grails.gorm.multiTenancy.tenantResolverClass' : SystemPropertyTenantResolver,
+                    'grails.mongodb.url'                           : "mongodb://${host}:${port}/tckMtDefaultDB" as String,
+                    'grails.mongodb.connections'                   : [
+                            'secondary': ['url': "mongodb://${host}:${port}/tckMtSecondaryDB" as String],
+                    ],
+            ]
+            multiTenantMultiDataSourceDatastore = new MongoDatastore(
+                    DatastoreUtils.createPropertyResolver(config), domainClasses
+            )
+        }
     }
 
     @Override
@@ -224,17 +213,34 @@ class GrailsDataMongoTckManager extends GrailsDataTckManager {
     @Override
     def getServiceForMultiTenantConnection(Class serviceType, String connectionName) {
         multiTenantMultiDataSourceDatastore
-                .getDatastoreForConnection(connectionName)
-                .getService(serviceType)
+                ?.getDatastoreForConnection(connectionName)
+                ?.getService(serviceType)
     }
 
     void setupValidator(Class entityClass, Validator validator = null) {
-        PersistentEntity entity = mappingContext.persistentEntities.find { PersistentEntity e -> e.javaClass == entityClass }
-        def messageSource = new StaticMessageSource()
-        def evaluator = new DefaultConstraintEvaluator(new DefaultConstraintRegistry(messageSource), mappingContext, Collections.emptyMap())
-        if (entity) {
-            mappingContext.addEntityValidator(entity, validator ?:
-                    new PersistentEntityValidator(entity, messageSource, evaluator))
+        if (mappingContext != null) {
+            PersistentEntity entity = mappingContext.persistentEntities.find { PersistentEntity e -> e.javaClass == entityClass }
+            def messageSource = new StaticMessageSource()
+            def evaluator = new DefaultConstraintEvaluator(new DefaultConstraintRegistry(messageSource), mappingContext, Collections.emptyMap())
+            if (entity) {
+                mappingContext.addEntityValidator(entity, validator ?:
+                        new PersistentEntityValidator(entity, messageSource, evaluator))
+            }
         }
+    }
+    static boolean isDockerAvailable() {
+        def candidates = [
+                System.getProperty('user.home') + '/.docker/run/docker.sock',
+                '/var/run/docker.sock',
+                System.getenv('DOCKER_HOST') ?: ''
+        ]
+        if (candidates.any { it && new File(it).exists() }) {
+            try {
+                return org.testcontainers.DockerClientFactory.instance().isDockerAvailable()
+            } catch (Throwable e) {
+                return false
+            }
+        }
+        return false
     }
 }
